test: integ test on CommonBitcoinLikeKeychains to fetch latest state before updating it
